backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
jinja2==3.1.4
weasyprint==63.0
python-dotenv==1.0.1
pydantic==2.10.2
email-validator==2.2.0
reportlab==4.2.5
